Update best practices with CSS guidelines
Including this because they are exactly the things we look out for during Code Review. Further, we point to this repo in the hiring code challenge and want to set applicants up for success by making these things explicit, as we have ruled people out for getting to many of these things wrong in the past.